rebase + fixes
